Add coap_debug.h and coap_udp.c
coap_debug.h allows to use a set of macros to add file, line and
function info to log messages.

coap_udp.c is the base for threaded serving coap messages.  It
creates a thread per dup sap to CoAP to allow registering of
service messages and to manage the retransmission and timeout
of CONfirmed messages.